+++ imports/experiment/server/index.js
@@ -1,23 +1,32 @@
 import { conditions } from "./game/conditions.js";
 import { init } from "./game/init.js";
-import {
+import callbacks from "./game/callbacks.js";
+
+const {
   onGameEnd,
   onGameStart,
   onRoundEnd,
   onRoundStart,
+  onSet,
+  onChange,
+  onAppend,
   onStageEnd,
   onStageStart
-} from "./game/callbacks.js";
+} = callbacks;
 
 import { bob } from "./bots.js";
+
 export const config = {
   conditions,
   init,
+  bots: [bob],
   onGameStart,
   onRoundStart,
   onStageStart,
+  onSet,
+  onAppend,
+  onChange,
   onStageEnd,
   onRoundEnd,
-  onGameEnd,
-  bots: [bob]
+  onGameEnd
 };