+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 - CSV Export includes the extracted Player's ID from the URL parameter (i.e., `playerIdParam` settings).
 - New onGameStart, onStageStart callbacks.
 - Simple page titles to help distinguish tabs.
+- New `append()` method on players, playerRounds, playerStages, games, stages and rounds.
+- Added `set()` to games, rounds and stages, and `get()` to games.
+- Added `onSet()`, `onAppend()` and `onChange()` callbacks.
 
 ### Changed
 
@@ -81,6 +84,9 @@ and this project adheres to
 - Loading issue between stages and at beginning of game.
 - Re-regestering of player with same playerID in lobby.
 - `.set()` value `0` was not supported, would be casted into `null`.
+- Callbacks marking the end of the game (including the end of the last round
+  and the last stage) were repeatedly called on a finished game until the whole
+  batch had ended.
 
 <!-- Add unreleased changes here -->